removed old initializer
